refactor: extract time formatting into util
